requirements.txt
fastapi==0.115.9
dapr-agents==0.3.1
dotenv==0.9.9
sentence-transformers==4.0.2
chromadb==1.0.4